Add quotientZero function
When you have an automaton after projection it can be reject acceptable
lines, but accept it if you add extra zeroes. So one of the solutions is
to go to all states and if you can reach one of final states, then you
add it into set of final states. quotientZero function do exactly it.
Also technical function to get length of used symbols into NFA